Correctly report solver CPU runtimes.
Previously, the solver*Time values were computed with
getrusage(RUSAGE_CHILDREN, ...). This function call returns resource usage
statistics for all descendants of the calling process (i.e., the solver
process and its descendants) that have terminated and have been waited
for. This means that subprocesses that have not called wait() are ignored.
Since this often underestimates the CPU times, we instead recompute the
CPU time for the watched process (current*Time) and add the time for
children that have already terminated (completed*Time).