Correctly report solver CPU runtimes. (#1)
We noticed that sometimes the solverCPUTime is way below the real CPU
usage of the solver (3s vs. 1800s). Unfortunately, this behaviour is not
deterministic. We believe that the problems stems from the following:

The solver*Time values are computed with getrusage(RUSAGE_CHILDREN, ...).
This function call returns resource usage statistics for all descendants
of the calling process (i.e., the solver process and its descendants) that
have terminated and have been waited for. This means that subprocesses
that have not called wait() are ignored.

After many hours of looking into this, the best fix that we
could come up with is to report the maximum over the measured CPU times.